fix a couple failing tests
git-svn-id: https://svn.participatoryculture.org/svn/channelguide2/trunk@1056 be7adf91-e322-0410-8f47-e6edb61c52aa

+++ channelguide/sessions/models.py
@@ -41,6 +41,7 @@ class Session(object):
         client.set(key, self, expires)
 
     def delete(self, connection):
-        client.delete(self._cache_key(self.session_key))
+        if self.session_key is not None:
+            client.delete(self._cache_key(self.session_key))
 
     delete_if_exists = delete
